refactor(ai): extract useAIChatFeedState and streamline stream token generator
