+++ frontend/src/routes/_layout/ai.tsx
@@ -10,12 +10,7 @@ import {
 import { DeleteThreadConfirmDialog } from "@/components/Chat/DeleteThreadConfirmDialog"
 import { PromptForm } from "@/components/Chat/PromptForm"
 import { RenameThreadDialog } from "@/components/Chat/RenameThreadDialog"
-import type {
-  AskUserAnswer,
-  AskUserToolPart,
-  ChatUIMessage,
-} from "@/components/Chat/types"
-import { useAIChatStream } from "@/hooks/useAIChatStream"
+import { useAIChatFeedState } from "@/hooks/useAIChatFeedState"
 
 export const Route = createFileRoute("/_layout/ai")({
   component: AIPage,
@@ -67,11 +62,8 @@ function filterAndSortThreads(
 
 function AIPage() {
   const queryClient = useQueryClient()
-  const { isStreaming, startStream, stop: stopStream } = useAIChatStream()
+  const promptInputRef = React.useRef<HTMLTextAreaElement>(null)
 
-  const [activeThreadId, setActiveThreadId] = React.useState<string | null>(
-    null,
-  )
   const [threadToRename, setThreadToRename] =
     React.useState<ChatThreadPublic | null>(null)
   const [threadToDelete, setThreadToDelete] =
@@ -85,8 +77,6 @@ function AIPage() {
   const [searchQuery, setSearchQuery] = React.useState("")
   const [sortOrder, setSortOrder] = React.useState<SortOption>("recent")
   const [isSortMenuOpen, setIsSortMenuOpen] = React.useState(false)
-
-  const promptInputRef = React.useRef<HTMLTextAreaElement>(null)
 
   const { data: modelsData } = useQuery({
     queryKey: ["ai-models"],
@@ -127,54 +117,25 @@ function AIPage() {
     }
   }, [modelsData, setSelectedModelId])
 
-  const [localMessages, setLocalMessages] = React.useState<ChatUIMessage[]>([])
-  const [pendingQuestion, setPendingQuestion] =
-    React.useState<AskUserToolPart | null>(null)
-
   const { data: threadsData, isLoading: isThreadsLoading } = useQuery({
     queryKey: ["ai-threads"],
     queryFn: () => AiThreadsService.listChatThreads({ skip: 0, limit: 100 }),
   })
 
   const threads = threadsData?.data ?? []
-  const initialLoadedRef = React.useRef(false)
 
-  React.useEffect(() => {
-    if (threads.length > 0 && !initialLoadedRef.current) {
-      setActiveThreadId(threads[0].id)
-      initialLoadedRef.current = true
-    }
-  }, [threads])
-
-  const { data: activeThreadDetail } = useQuery({
-    queryKey: ["ai-thread", activeThreadId],
-    queryFn: () => AiThreadsService.getChatThread({ id: activeThreadId! }),
-    enabled: !!activeThreadId,
-  })
-
-  React.useEffect(() => {
-    if (!isStreaming) {
-      if (activeThreadId && activeThreadDetail?.transcript) {
-        const msgs =
-          (activeThreadDetail.transcript as { messages?: ChatUIMessage[] })
-            ?.messages || []
-        setLocalMessages(msgs)
-      } else if (!activeThreadId) {
-        setLocalMessages([])
-      }
-    }
-  }, [activeThreadDetail, isStreaming, activeThreadId])
-
-  const createThreadMutation = useMutation({
-    mutationFn: (prompt?: string) =>
-      AiThreadsService.createChatThread({
-        requestBody: { origin: "composer", prompt },
-      }),
-    onSuccess: (newThread) => {
-      queryClient.invalidateQueries({ queryKey: ["ai-threads"] })
-      setActiveThreadId(newThread.id)
-    },
-  })
+  const {
+    activeThreadId,
+    setActiveThreadId,
+    localMessages,
+    setLocalMessages,
+    pendingQuestion,
+    isStreaming,
+    stopStream,
+    handleSendMessage,
+    handleQuestionAnswer,
+    handleNewChat,
+  } = useAIChatFeedState({ threads, selectedModelId })
 
   const updateThreadMutation = useMutation({
     mutationFn: ({
@@ -220,102 +181,8 @@ function AIPage() {
     }
   }, [openMenuThreadId, isSortMenuOpen])
 
-  function handleSendMessage(text: string) {
-    if (!text.trim() || isStreaming) return
-
-    const userMsg: ChatUIMessage = {
-      id: `local-user-${Date.now()}`,
-      role: "user",
-      parts: [{ type: "text", text: text.trim() }],
-      createdAt: new Date().toISOString(),
-    }
-
-    const assistantMsgId = `local-assistant-${Date.now()}`
-    const assistantMsg: ChatUIMessage = {
-      id: assistantMsgId,
-      role: "assistant",
-      parts: [],
-      createdAt: new Date().toISOString(),
-    }
-
-    setLocalMessages((prev) => [...prev, userMsg, assistantMsg])
-    setPendingQuestion(null)
-
-    async function execute() {
-      let targetThreadId = activeThreadId
-      if (!targetThreadId) {
-        try {
-          const newThread = await createThreadMutation.mutateAsync(text.trim())
-          targetThreadId = newThread.id
-          setActiveThreadId(newThread.id)
-        } catch {
-          return
-        }
-      }
-
-      startStream(
-        targetThreadId,
-        text.trim(),
-        {
-          onThought: (content) => {
-            setLocalMessages((prev) =>
-              prev.map((msg) => {
-                if (msg.id !== assistantMsgId) return msg
-                const parts = [...msg.parts]
-                const lastPart = parts[parts.length - 1]
-                if (lastPart && lastPart.type === "thought") {
-                  lastPart.content += content
-                } else {
-                  parts.push({ type: "thought", content })
-                }
-                return { ...msg, parts }
-              }),
-            )
-          },
-          onTextDelta: (delta) => {
-            setLocalMessages((prev) =>
-              prev.map((msg) => {
-                if (msg.id !== assistantMsgId) return msg
-                const parts = [...msg.parts]
-                const lastPart = parts[parts.length - 1]
-                if (lastPart && lastPart.type === "text") {
-                  lastPart.text += delta
-                } else {
-                  parts.push({ type: "text", text: delta })
-                }
-                return { ...msg, parts }
-              }),
-            )
-          },
-          onDone: () => {
-            queryClient.invalidateQueries({ queryKey: ["ai-threads"] })
-            queryClient.invalidateQueries({
-              queryKey: ["ai-thread", targetThreadId],
-            })
-          },
-        },
-        selectedModelId,
-      )
-    }
-
-    execute()
-  }
-
-  function handleQuestionAnswer(_toolCallId: string, answers: AskUserAnswer[]) {
-    const text = answers
-      .map((a) => a.answer)
-      .filter(Boolean)
-      .join("; ")
-    if (text) {
-      handleSendMessage(text)
-    }
-  }
-
-  function handleNewChat() {
-    stopStream()
-    setActiveThreadId(null)
-    setLocalMessages([])
-    setPendingQuestion(null)
+  function onNewChatClick() {
+    handleNewChat()
     setTimeout(() => promptInputRef.current?.focus(), 50)
   }
 
@@ -424,7 +291,7 @@ function AIPage() {
           },
           onToggleRecents: () => setRecentsOpen((prev) => !prev),
           onToggleArchived: () => setArchivedOpen((prev) => !prev),
-          onNewChat: handleNewChat,
+          onNewChat: onNewChatClick,
         }}
         actions={{
           onSelect: (threadId) => {
